rename: tree -> category-tree
