+++ src/popup/category-tree.mjs
